perf: GPU-less static bottleneck analysis (AOT sm_90a + ptxas) drives kernel tuning
- benchmarks/static_analysis.py: AOT-compile all 9 Triton kernels for
  sm_90a with pip-installed ptxas; report regs/spills/smem, theoretical
  GH100 occupancy, and wave count of the real Mixtral-shape grids
- fix op4 verify kernel int32/int64 loop-carried type drift: compiles
  fine in interpreter mode but FAILS real sm_90 compilation (caught by
  AOT compile, invisible to numpy interpreter)
- tuning from ptxas sweeps:
  * Kernel A num_warps 4->32: regs 255->64, spill 7.6KB->0.9KB/thread
    (the O((2N)^2) rank matrix was spilling to local memory)
  * GEMM1 BK 128->64: regs 232->168, occupancy 12%->19%
  * GEMM2 BN 128->64: regs 150->96, occupancy 19%->31%, grid 2x
- benchmarks/roofline.py: exact byte-accounting speedup projection;
  shows expert budget is decisive: B=8 caps at 1.17x even with MAT=4.5,
  B=2-3 projects 2-3x on H200 (MAT sweep, Phase-0 will pin the curve)
- regression: 56 CPU + 17 interpreter tests green

+++ treemoe/kernels/op4_commit.py
@@ -110,7 +110,13 @@ if HAS_TRITON:
             start = tl.load(child_start_ptr + node)
             accepted = -1
             for j in range(0, n_kids):
-                c = tl.load(child_list_ptr + start + j)
+                # .to(int32): keep the loop-carried `accepted`/`node` scalars at a
+                # fixed int32 type — mixing the i64 load into tl.where would flip
+                # the carried type and break the sm_90 compile (found by AOT
+                # compile in benchmarks/static_analysis.py; interpreter mode is
+                # too permissive to catch it). Node ids < 2N=128, int32 is safe;
+                # tl.store casts back to the i64 buffers.
+                c = tl.load(child_list_ptr + start + j).to(tl.int32)
                 tok = tl.load(tree_tokens_ptr + c)
                 hit = (tok == target_top) & (accepted < 0)
                 accepted = tl.where(hit, c, accepted)
